feat(S13P21D204-177): 웨어러블데이터 기반 AI서버에서 response반환. -fcm토큰 register/unregister api와  헬스데이터 이상탐지 기능 user server, health server로 분리

+++ helloworld-server/healthserver/src/main/java/com/example/helloworld/healthserver/dto/HealthDtos.java
@@ -11,7 +11,7 @@ public final class HealthDtos {
     // 생성 요청 (전부 nullable 허용)
     public record CreateRequest(
             Instant date,
-            Integer stress,
+            Double stress,
             Integer heartrate
     ) {}
 
@@ -19,7 +19,7 @@ public final class HealthDtos {
     public record GetResponse(
             @JsonProperty("health_id") Long healthId,
             Instant date,
-            Integer stress,
+            Double stress,
             Integer heartrate
     ) {}
